Remove the big.Int mistake
int64 is big enough. No need for the big.Int pain.

+++ 2019/days/q13_test.go
@@ -38,7 +38,7 @@ var _ = Describe("Q13", func() {
 	It("does part B", func() {
 		g.Pay()
 		score := g.Run()
-		Expect(score).To(Equal(19297))
+		Expect(score).To(Equal(int64(19297)))
 	})
 
 })
